chore: rename statstat.rs to stat.rs
