+++ lib/src/edhoc.rs
@@ -2,13 +2,13 @@ use crate::InitiatorIdentity;
 use digest::Digest;
 use lakers_shared::{Crypto as CryptoTrait, *};
 mod psk;
-mod statstat;
+mod stat;
 
 use psk::{
     i_parse_message_2_psk, i_prepare_message_3_psk, i_verify_message_2_psk, r_parse_message_3_psk,
     r_parse_message_3_psk_with_cred_resolver, r_prepare_message_2_psk, r_verify_message_3_psk,
 };
-use statstat::{
+use stat::{
     i_parse_message_2_statstat, i_prepare_message_3_statstat, i_verify_message_2_statstat,
     r_parse_message_3_statstat, r_prepare_message_2_statstat, r_verify_message_3_statstat,
 };
